UPSTREAM: time: move the timecounter/cyclecounter code into its own file.
(cherry picked from commit 74d23cc704d19732e70ef1579a669f7d5f09dd9a)

The timecounter code has almost nothing to do with the clocksource
code. Let it live in its own file. This will help isolate the
timecounter users from the clocksource users in the source tree.

Acked-by: Jeff Kirsher <jeffrey.t.kirsher@intel.com>
